exercice1/requirements.txt
asgiref==3.8.1
black==25.1.0
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.2.1
Django==5.2.1
djangorestframework==3.16.0
flake8==7.2.0
idna==3.10
isort==6.0.1
mccabe==0.7.0
mypy_extensions==1.1.0
numpy==2.2.6
packaging==25.0
pathspec==0.12.1
platformdirs==4.3.8
pycodestyle==2.13.0
pyflakes==3.3.2
requests==2.32.3
sqlparse==0.5.3
urllib3==2.4.0
